Define a dApp's host once, in shared
`getDappHost` in shared re-implemented
`normalizeHost(getHostFromUrl(...))` from the mobile ecosystemUtils. The
latter is what the authorization path uses - it is how
`isDappHostVerified` and the WebView's senderHost are derived - so
analytics was attributing transactions to a dApp identity computed by a
second, independent copy of the rule, free to drift from the one the
wallet actually authorizes against.

Move both helpers into shared and have `getDappHost` compose them. It
stays lenient about a missing scheme, because analytics receives bare
hosts from the in-app browser, while `getHostFromUrl` stays strict,
because the authorization path must not accept something that is not a
real URL. Test that both paths agree.
