+++ packages/shared/src/utils/index.ts
@@ -5,3 +5,4 @@ export * from '../utils/types'
 export * from '../utils/network'
 export * from '../utils/errors'
 export * from '../utils/objects'
+export * from '../utils/dApps'
